sclang: Refactor tag enum to PyrTag
Change tag enum to a proper named and strongly-typed enum, PyrTag. This increases debugger visibility of tags, and makes tag<>int conversions explicit. This commit should entail no functional change - in a few places where implicit type conversions were being done, there is now a static_cast, but the behavior should be identical to what it was before. This also converts a few explicit tag comparisons to e.g. "IsInt" form.

+++ lang/LangSource/PyrObject.h
@@ -246,7 +246,7 @@ inline struct PyrClass* classOfSlot(PyrSlot *slot)
 	PyrClass *classobj;
 	int tag;
 	if (IsFloat(slot)) classobj = class_float;
-	else if ((tag = GetTag(slot) & 0xF) == 1) classobj = slotRawObject(slot)->classptr;
+	else if ((tag = static_cast<int>(GetTag(slot)) & 0xF) == 1) classobj = slotRawObject(slot)->classptr;
 	else classobj = gTagClassTable[tag];
 
 	return classobj;
